Version 2.0alpha
An alpha release of a complete redesign of the plug-in code. It should be much faster than 1.0.6.1, and about as fast as the 1.0.7 version that I just uploaded.

The API of this plug-in is completely different than before.

Consult the help file in the ZIP to see the details.

This is very much a work in progress! I use my own plug-in mainly for work, and since it's summer I haven't had the chance for some serious dogfooding. In summary I added some functionality that I was missing in the first version, and existing functionality should be more accessible.

Some highlights:

* More fine-grained filtering: gather results that match one or more alternatives (OR-semantics), if these lines also match zero or more other terms (AND-semantics), until they match zero or more even other terms (AND-NOT).
There's no real need any more to filter the results again, you can just use AND.
I think weeding out results using AND-NOT can be very useful too, for example you can quickly grep for all logging of a certain component, and quickly remove debug-level logging. This was not possible before (without using fragile tricks as I've been doing myself).

* Switch between highlighted line-numbered view (the default) and a "raw" view where you just see the plain text results without any line numbering. (Note that this disable pretty much all of the functionality of the plug-in as well.)
Nice for yanking text, for example when writing a problem report where the original line numbers are irrelevant.

* Search for an extra pattern (existing functionality), but not just N lines ahead. You can now also specify a stop-pattern. So you can now look for lines matching M if they match pattern E(xtra) before pattern S(top). The default stop pattern is the empty line.
I think this will be useful when filtering protocol log files that contain decoded messages separated by a blank line.